Remove None declump_method
Originally, the None declump method would pass the thresholded image to
the skimage watershed function, which led to a different segmentation
than if the distance transform or intensity image were passed.

However, technically that's not really `None` declumping and if it was,
it's behaving differently from IDPrimary, which just returns a labelled
array.

This comit reverts just labelled the thresholded image because it begins
to get strange when you consider a marker-based watershed. In the GUI,
you can provide a marker image, but if you select None declumping it
is entirely ignored.

TLDR: Watershed should be for watershed, not only scipy.ndimage.label