Remove intensity-image based maxima finding in watershed
Finding maxima in an intensity image was originally added in this
refactor, but it doesn't make sense alongside the `markers` method.

Additionally, having an intensity-based maxima method uses the same
intensity image as intensity declumping, which also doesn't make sense.

A user should instead calculate their own markers from an intensity image
using the FindMaxima module, which can then be passed in to watershed
as markers.